// src/app/modules/shared/models/cloud.ts
export interface CloudConfig {
    ferrumCloudId: string;
    ferrumCloudUrl: string;
    ferrumCloudToken: string;
    ferrumCloudIp: string;
    ferrumCloudPort: string;
    redisPass: string;
    redisIntelPass: string;
    encryptKey: string;
    esUser: string;
    esPass: string;
    esIntelUser: string;
    esIntelPass: string;
    clusterNodePublicKey: string;
}